# microservices/api-gateway/openapi/fragments/user-service.yaml
openapi: 3.1.0
info:
  title: User Service API
  description: User management microservice
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:9001
    description: User Service (direct)
  - url: http://localhost:9000/api
    description: API Gateway
...
paths:
...
  /api/users:
    get:
      tags:
        - users
      summary: Get all users
      description: Retrieve all users with optional limit
      operationId: getAllUsers
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          $ref: '#/components/responses/UsersSuccessResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - users
      summary: Create a new user
      description: Create a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          $ref: '#/components/responses/UserCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/verify:
    get:
      tags:
        - authentication
      summary: Verify JWT token
      description: Verify if the provided JWT token is valid and return user information
      operationId: verifyToken
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/VerifyTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/refresh:
    post:
      tags:
        - authentication
      summary: Refresh JWT token
      description: Generate a new JWT token from an existing valid token
      operationId: refreshToken
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/LoginSuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/change-password:
    put:
      tags:
        - password-management
      summary: Change password
      description: |
        Change user password. Requires old password verification.
        After successful password change, the current token is blacklisted and user must login again.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            examples:
              validRequest:
                summary: Valid password change request
                value:
                  oldPassword: OldPassword123!
                  newPassword: NewPassword123!
      responses:
        '200':
          $ref: '#/components/responses/ChangePasswordSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/forgot-password:
    post:
      tags:
        - password-management
      summary: Request password reset
      description: |
        Request a password reset link. If the email exists in the system, a reset token will be generated
        and sent via email. Returns generic success message for security reasons.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            examples:
              withDefaultUrl:
                summary: With default reset URL
                value:
                  email: admin@gmail.com
              withCustomUrl:
                summary: With custom reset URL
                value:
                  email: admin@gmail.com
                  resetUrl: https://myapp.com/reset-password
      responses:
        '200':
          $ref: '#/components/responses/ForgotPasswordSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/reset-password:
    post:
      tags:
        - password-management
      summary: Reset password with token
      description: |
        Reset password using the token received via email. The token is valid for 1 hour
        and can only be used once.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            examples:
              validRequest:
                summary: Valid reset request
                value:
                  token: fFzw791jCP_vfVFn_NNj5DF68pw9egrkbeKTTIWaW-c
                  newPassword: NewPassword123!
      responses:
        '200':
          $ref: '#/components/responses/ResetPasswordSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: User ID
          examples:
            - '1'
        email:
          type: string
          format: email
          description: User email address
          examples:
            - john.doe@example.com
        name:
          type: string
          description: User full name
          examples:
            - John Doe
        birthdate:
          oneOf:
            - type: string
              format: date
            - type: 'null'
          description: User birthdate
          examples:
            - '1990-01-15'
        status:
          type: integer
          description: User status (0 = Inactive, 1 = Active)
          enum:
            - 0
            - 1
          examples:
            - 1
        lastLoginAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Last login timestamp
          examples:
            - '2024-01-15T10:30:00'
        verifiedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Email verification timestamp
          examples:
            - '2024-01-15T10:30:00'
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          examples:
            - '2024-01-15T10:30:00'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          examples:
            - '2024-01-15T10:30:00'
    CreateUserRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          description: User email address
          examples:
            - john.doe@example.com
        name:
          type: string
          description: User full name
          minLength: 1
          examples:
            - John Doe
        password:
          oneOf:
            - type: string
              format: password
            - type: 'null'
          description: User password (optional)
          examples:
            - SecurePass123!
        birthdate:
          oneOf:
            - type: string
              format: date
            - type: 'null'
          description: User birthdate (optional)
          examples:
            - '1990-01-15'
...
    UserSuccessResponse:
      type: object
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          examples:
            - true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
          examples:
            - User retrieved successfully
    UsersSuccessResponse:
      type: object
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          examples:
            - true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        message:
          type: string
          examples:
            - Users retrieved successfully
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          examples:
            - false
        error:
          type: string
          description: Error message
          examples:
            - User not found
        code:
          type: string
          description: Error code
          enum:
            - VALIDATION_ERROR
            - NOT_FOUND
            - INVALID_OPERATION
            - DUPLICATE_ERROR
            - INTERNAL_ERROR
            - USER_NOT_FOUND
          examples:
            - USER_NOT_FOUND
...
    LoginData:
      type: object
      required:
        - user
        - token
        - expiresAt
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT access token
          examples:
            - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZW1haWwiOiJhZG1pbkBnbWFpbC5jb20iLCJpYXQiOjE3NjA0OTM4ODIsImV4cCI6MTc2MDU4MDI4Mn0.JoVgK8xIwBniSMIyr8S4wl4-yC86VKJBfixa-hIs2WQ
        expiresAt:
          type: string
          format: date-time
          description: Token expiration timestamp
          examples:
            - '2025-10-16T02:04:42.036492423'
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          format: password
          description: Current password
          minLength: 6
          examples:
            - OldPassword123!
        newPassword:
          type: string
          format: password
          description: New password
          minLength: 6
          examples:
            - NewPassword123!
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email address
          examples:
            - admin@gmail.com
        resetUrl:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
          description: Custom reset URL (optional, defaults to http://localhost:3000/reset-password)
          examples:
            - https://myapp.com/reset-password
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token received via email
          examples:
            - fFzw791jCP_vfVFn_NNj5DF68pw9egrkbeKTTIWaW-c
        newPassword:
          type: string
          format: password
          description: New password
          minLength: 6
          examples:
            - NewPassword123!
  responses:
...
    LoginSuccessResponse:
      description: Login successful
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - data
              - message
            properties:
              success:
                type: boolean
                examples:
                  - true
              data:
                $ref: '#/components/schemas/LoginData'
              message:
                type: string
                examples:
                  - Login successful
          examples:
            success:
              summary: Successful login
              value:
                success: true
                data:
                  user:
                    id: 1
                    email: admin@gmail.com
                    name: Admin User
                    birthdate: '1990-01-01'
                    status: 1
                    lastLoginAt: '2025-10-15T02:04:42.046852'
                    createdAt: '2025-10-14T04:03:07.874711'
                    updatedAt: '2025-10-15T02:04:42.057310'
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expiresAt: '2025-10-16T02:04:42.036492423'
                message: Login successful
    VerifyTokenResponse:
      description: Token is valid
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - data
              - message
            properties:
              success:
                type: boolean
                examples:
                  - true
              data:
                $ref: '#/components/schemas/User'
              message:
                type: string
                examples:
                  - Token is valid
...
    ChangePasswordSuccessResponse:
      description: Password changed successfully
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - message
            properties:
              success:
                type: boolean
                examples:
                  - true
              message:
                type: string
                examples:
                  - Password changed successfully. Please login again.
    ForgotPasswordSuccessResponse:
      description: Password reset request processed
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - message
            properties:
              success:
                type: boolean
                examples:
                  - true
              message:
                type: string
                examples:
                  - If your email exists in our system, you will receive a password reset link
    ResetPasswordSuccessResponse:
      description: Password reset successfully
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - message
            properties:
              success:
                type: boolean
                examples:
                  - true
              message:
                type: string
                examples:
                  - Password reset successfully. You can now login with your new password.
...
    UsersSuccessResponse:
      description: Users retrieved successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UsersSuccessResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UserCreatedResponse:
      description: User created successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserSuccessResponse'
    BadRequestError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnauthorizedError:
      description: Authentication failed or invalid credentials
      content:
        application/json:
          schema:
            type: object
            required:
              - success
              - message
            properties:
              success:
                type: boolean
                examples:
                  - false
              message:
                type: string
                examples:
                  - Invalid email or password
          examples:
            invalidCredentials:
              summary: Invalid credentials
              value:
                success: false
                message: Invalid email or password
            invalidToken:
              summary: Invalid token
              value:
                success: false
                message: Invalid or expired token
            missingToken:
              summary: Missing authorization header
              value:
                success: false
                message: Missing or invalid Authorization header
  parameters:
...
    LimitQuery:
      name: limit
      in: query
      description: Maximum number of users to return
      required: false
      schema:
        type: integer
        default: 10000
        minimum: 1
        maximum: 10000
...